🛠️ refactor: 여러 서비스 리팩토링 및 수정 (팀 변경 로직, API 명세 수정, H2 설정 등)

+++ src/main/java/prography/spring/pingpong/domain/game/model/dto/GameStartRequestDto.java
@@ -1,4 +1,4 @@
 package prography.spring.pingpong.domain.game.model.dto;
 
-public record GameStartRequestDto(Long userId) {
+public record GameStartRequestDto(Integer userId) {
 }
