+++ src/main/java/prography/spring/pingpong/domain/room/model/dto/RoomCreateRequestDto.java
@@ -1,4 +1,4 @@
 package prography.spring.pingpong.domain.room.model.dto;
 
-public record RoomCreateRequestDto(Long userId, String roomType, String title) {
+public record RoomCreateRequestDto(int userId, String roomType, String title) {
 }
