+++ src/main/java/prography/spring/pingpong/domain/userroom/model/dto/TeamChangeRequestDto.java
@@ -1,4 +1,4 @@
 package prography.spring.pingpong.domain.userroom.model.dto;
 
-public record TeamChangeRequestDto(Long userId) {
+public record TeamChangeRequestDto(int userId) {
 }
